% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{BottomOutlierIdentifier}
\alias{BottomOutlierIdentifier}
\title{Bottom outlier identifier.}
\usage{
BottomOutlierIdentifier(dataVec, lowerAndUpperThresholds)
}
\arguments{
\item{dataVec}{data vector}

\item{lowerAndUpperThresholds}{outlier identifier parameters}
}
\description{
Find the bottom outliers for a data vector.
}
