% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{BottomOutlierPosition}
\alias{BottomOutlierPosition}
\title{Bottom outlier positions finder.}
\usage{
BottomOutlierPosition(dataVec,
  outlierIdentifier = HampelIdentifierParameters)
}
\arguments{
\item{dataVec}{A data vector.}

\item{outlierIdentifier}{An outlier identifier function.}
}
\description{
Find the bottom outlier positions in a data vector.
}
