% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{BottomOutliersOnlyThresholds}
\alias{BottomOutliersOnlyThresholds}
\title{Bottom outliers only thresholds}
\usage{
BottomOutliersOnlyThresholds(pair)
}
\arguments{
\item{pair}{A pair of bottom and top thresholds.}
}
\description{
Drops the bottom threshold of a pair of thresholds.
}
\seealso{
Other Outlier identifier parameters: 
\code{\link{HampelIdentifierParameters}()},
\code{\link{QuartileIdentifierParameters}()},
\code{\link{SPLUSQuartileIdentifierParameters}()},
\code{\link{TopOutliersOnlyThresholds}()}
}
\concept{Outlier identifier parameters}
