% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{HampelIdentifierParameters}
\alias{HampelIdentifierParameters}
\title{Hampel identifier parameters.}
\usage{
HampelIdentifierParameters(dataVec)
}
\arguments{
\item{dataVec}{A data vector.}
}
\description{
Find an Hampel outlier threshold for a data vector.
}
