% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{OutlierIdentifier}
\alias{OutlierIdentifier}
\title{Outlier identifier.}
\usage{
OutlierIdentifier(dataVec, lowerAndUpperThresholds)
}
\arguments{
\item{dataVec}{A data vector.}

\item{lowerAndUpperThresholds}{outlier identifier parameters}
}
\description{
Find an outlier threshold for a data vector.
}
