% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{SPLUSQuartileIdentifierParameters}
\alias{SPLUSQuartileIdentifierParameters}
\title{SPLUS quartile identifier parameters}
\usage{
SPLUSQuartileIdentifierParameters(dataVec)
}
\arguments{
\item{dataVec}{A data vector.}
}
\description{
Find an SPLUS Quartile outlier for a data vector.
}
