% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{TopOutlierIdentifier}
\alias{TopOutlierIdentifier}
\title{Top outlier identifier.}
\usage{
TopOutlierIdentifier(data, identifier,
  lowerAndUpperThresholds = identifier(data))
}
\arguments{
\item{data}{A data vector.}

\item{identifier}{An outlier identifier function.}

\item{lowerAndUpperThresholds}{Outlier identifier parameters.}
}
\description{
Find the top outliers for a data vector
}
