% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{TopOutliersOnlyThresholds}
\alias{TopOutliersOnlyThresholds}
\title{Top outliers only thresholds}
\usage{
TopOutliersOnlyThresholds(pair)
}
\arguments{
\item{pair}{A pair of bottom and top thresholds.}
}
\description{
Drops the bottom threshold of a pair of thresholds.
}
\seealso{
Other Outlier identifier parameters: 
\code{\link{BottomOutliersOnlyThresholds}()},
\code{\link{HampelIdentifierParameters}()},
\code{\link{QuartileIdentifierParameters}()},
\code{\link{SPLUSQuartileIdentifierParameters}()}
}
\concept{Outlier identifier parameters}
