% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{BottomOutlierIdentifier}
\alias{BottomOutlierIdentifier}
\title{Bottom outlier identifier.}
\usage{
BottomOutlierIdentifier(
  data,
  identifier,
  lowerAndUpperThresholds = identifier(data),
  valueQ = FALSE
)
}
\arguments{
\item{data}{data vector}

\item{identifier}{An outlier identifier function.}

\item{lowerAndUpperThresholds}{Outlier identifier parameters.}

\item{valueQ}{Should values be returned or not?}
}
\value{
A numeric vector of outliers or a logical vector.
}
\description{
Find the bottom outliers for a data vector.
}
\seealso{
Other Outlier identifiers: 
\code{\link{BottomOutlierPosition}()},
\code{\link{OutlierIdentifier}()},
\code{\link{OutlierPosition}()},
\code{\link{TopOutlierIdentifier}()},
\code{\link{TopOutlierPosition}()}
}
\concept{Outlier identifiers}
