% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{BottomOutlierIdentifier}
\alias{BottomOutlierIdentifier}
\title{Bottom outlier identifier.}
\usage{
BottomOutlierIdentifier(data, identifier,
  lowerAndUpperThresholds = identifier(data))
}
\arguments{
\item{data}{data vector}

\item{identifier}{An outlier identifier function.}

\item{lowerAndUpperThresholds}{Outlier identifier parameters.}
}
\description{
Find the bottom outliers for a data vector.
}
