% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{OutlierIdentifier}
\alias{OutlierIdentifier}
\title{Outlier identifier.}
\usage{
OutlierIdentifier(data, identifier,
  lowerAndUpperThresholds = identifier(data))
}
\arguments{
\item{data}{A data vector.}

\item{identifier}{An outlier identifier function.}

\item{lowerAndUpperThresholds}{Outlier identifier parameters.}
}
\value{
A numeric vector of outliers.
}
\description{
Find an outlier threshold for a data vector.
}
\details{
The outlier identifier function \code{identifier}
should return a list or tuple of two numbers \code{c(lowerThreshold, upperThreshold)}.
}
