% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{OutlierIdentifier}
\alias{OutlierIdentifier}
\title{Outlier identifier.}
\usage{
OutlierIdentifier(
  data,
  identifier,
  lowerAndUpperThresholds = identifier(data),
  valueQ = FALSE
)
}
\arguments{
\item{data}{A data vector.}

\item{identifier}{An outlier identifier function.}

\item{lowerAndUpperThresholds}{Outlier identifier parameters.}

\item{valueQ}{Should values be returned or not?}
}
\value{
A numeric vector of outliers or a logical vector.
}
\description{
Find an outlier threshold for a data vector.
}
\details{
The outlier identifier function \code{identifier}
should return a list or tuple of two numbers \code{c(lowerThreshold, upperThreshold)}.
}
\seealso{
Other Outlier identifiers: 
\code{\link{BottomOutlierIdentifier}()},
\code{\link{BottomOutlierPosition}()},
\code{\link{OutlierPosition}()},
\code{\link{TopOutlierIdentifier}()},
\code{\link{TopOutlierPosition}()}
}
\concept{Outlier identifiers}
