% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{OutlierPosition}
\alias{OutlierPosition}
\title{Outlier positions finder.}
\usage{
OutlierPosition(dataVec, outlierIdentifier = HampelIdentifierParameters)
}
\arguments{
\item{dataVec}{A data vector.}

\item{outlierIdentifier}{An outlier identifier function.}
}
\description{
Find the outlier positions in a data vector.
}
