% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{QuartileIdentifierParameters}
\alias{QuartileIdentifierParameters}
\title{Quartile identifier parameters}
\usage{
QuartileIdentifierParameters(data)
}
\arguments{
\item{data}{A data vector.}
}
\description{
Find an Quartile outlier for a data vector.
}
