% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{TopOutlierIdentifier}
\alias{TopOutlierIdentifier}
\title{Top outlier identifier.}
\usage{
TopOutlierIdentifier(
  data,
  identifier,
  lowerAndUpperThresholds = identifier(data),
  valueQ = FALSE
)
}
\arguments{
\item{data}{A data vector.}

\item{identifier}{An outlier identifier function.}

\item{lowerAndUpperThresholds}{Outlier identifier parameters.}

\item{valueQ}{Should values be returned or not?}
}
\value{
A numeric vector of outliers or a logical vector.
}
\description{
Find the top outliers for a data vector
}
\seealso{
Other Outlier identifiers: 
\code{\link{BottomOutlierIdentifier}()},
\code{\link{BottomOutlierPosition}()},
\code{\link{OutlierIdentifier}()},
\code{\link{OutlierPosition}()},
\code{\link{TopOutlierPosition}()}
}
\concept{Outlier identifiers}
