% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{TopOutlierIdentifier}
\alias{TopOutlierIdentifier}
\title{Top outlier identifier.}
\usage{
TopOutlierIdentifier(dataVec, lowerAndUpperThresholds)
}
\arguments{
\item{dataVec}{A data vector.}

\item{lowerAndUpperThresholds}{outlier identifier parameters}
}
\description{
Find the top outliers for a data vector
}
