% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierIdentifiers.R
\name{TopOutlierPosition}
\alias{TopOutlierPosition}
\title{Top outlier positions finder.}
\usage{
TopOutlierPosition(
  data,
  identifier = SPLUSQuartileIdentifierParameters,
  lowerAndUpperThresholds = identifier(data)
)
}
\arguments{
\item{data}{A data vector.}

\item{identifier}{An outlier identifier function.}

\item{lowerAndUpperThresholds}{Outlier identifier parameters.}
}
\description{
Find the top outlier positions in a data vector.
}
\seealso{
Other Outlier identifiers: 
\code{\link{BottomOutlierIdentifier}()},
\code{\link{BottomOutlierPosition}()},
\code{\link{OutlierIdentifier}()},
\code{\link{OutlierPosition}()},
\code{\link{TopOutlierIdentifier}()}
}
\concept{Outlier identifiers}
